# Use ps.Map for Counters and Sets, remove Metadata in favour of Latest.