--- a/experimental/demoprobe/main.go
+++ b/experimental/demoprobe/main.go
@@ -90,14 +90,14 @@ func demoReport(nodeCount int) report.Report {
 		)
 
 		// Endpoint topology
-		r.Endpoint = r.Endpoint.AddNode(srcPortID, report.MakeNode().WithMetadata(map[string]string{
+		r.Endpoint = r.Endpoint.AddNode(srcPortID, report.MakeNode().WithLatests(map[string]string{
 			process.PID: "4000",
 			"name":      c.srcProc,
 			"domain":    "node-" + src,
 		}).WithEdge(dstPortID, report.EdgeMetadata{
 			MaxConnCountTCP: newu64(uint64(rand.Intn(100) + 10)),
 		}))
-		r.Endpoint = r.Endpoint.AddNode(dstPortID, report.MakeNode().WithMetadata(map[string]string{
+		r.Endpoint = r.Endpoint.AddNode(dstPortID, report.MakeNode().WithLatests(map[string]string{
 			process.PID: "4000",
 			"name":      c.dstProc,
 			"domain":    "node-" + dst,
@@ -106,10 +106,10 @@ func demoReport(nodeCount int) report.Report {
 		}))
 
 		// Address topology
-		r.Address = r.Address.AddNode(srcAddressID, report.MakeNode().WithMetadata(map[string]string{
+		r.Address = r.Address.AddNode(srcAddressID, report.MakeNode().WithLatests(map[string]string{
 			docker.Name: src,
 		}).WithAdjacent(dstAddressID))
-		r.Address = r.Address.AddNode(srcAddressID, report.MakeNode().WithMetadata(map[string]string{
+		r.Address = r.Address.AddNode(srcAddressID, report.MakeNode().WithLatests(map[string]string{
 			docker.Name: dst,
 		}).WithAdjacent(srcAddressID))