--- a/probe/host/tagger.go
+++ b/probe/host/tagger.go
@@ -31,16 +31,14 @@ func (t Tagger) Tag(r report.Report) (report.Report, error) {
 			report.HostNodeID: t.hostNodeID,
 			report.ProbeID:    t.probeID,
 		}
-		parents = report.Sets{
-			report.Host: report.MakeStringSet(t.hostNodeID),
-		}
+		parents = report.EmptySets.Add(report.Host, report.MakeStringSet(t.hostNodeID))
 	)
 
 	// Explicity don't tag Endpoints and Addresses - These topologies include pseudo nodes,
 	// and as such do their own host tagging
 	for _, topology := range []report.Topology{r.Process, r.Container, r.ContainerImage, r.Host, r.Overlay} {
 		for id, node := range topology.Nodes {
-			topology.AddNode(id, node.WithMetadata(metadata).WithParents(parents))
+			topology.AddNode(id, node.WithLatests(metadata).WithParents(parents))
 		}
 	}
 	return r, nil