Part of #502, helps towards #812 & #854.