@tomwilkie 2016-01-19
This only shows a small improvement vs master.

This change:
```
Toms-MacBook-Pro:render twilkie$ go test -bench .
PASS
BenchmarkEndpointRender-8                	   10000	    120894 ns/op	   46217 B/op	     655 allocs/op
BenchmarkEndpointStats-8                 	100000000	        18.9 ns/op       0 B/op	       0 allocs/op
BenchmarkProcessRender-8                 	   10000	    229275 ns/op	   87324 B/op	    1166 allocs/op
BenchmarkProcessStats-8                  	20000000	        64.5 ns/op       0 B/op	       0 allocs/op
BenchmarkProcessWithContainerNameRender-8	    5000	    259938 ns/op	   96385 B/op	    1310 allocs/op
BenchmarkProcessWithContainerNameStats-8 	20000000	        75.3 ns/op       0 B/op	       0 allocs/op
BenchmarkProcessNameRender-8             	    5000	    289304 ns/op	  110414 B/op	    1414 allocs/op
BenchmarkProcessNameStats-8              	100000000	        20.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerRender-8               	    3000	    426344 ns/op	  149200 B/op	    2037 allocs/op
BenchmarkContainerStats-8                	    3000	    355616 ns/op	  128917 B/op	    1825 allocs/op
BenchmarkContainerWithImageNameRender-8  	    3000	    435100 ns/op	  160829 B/op	    2170 allocs/op
BenchmarkContainerWithImageNameStats-8   	    3000	    350997 ns/op	  128936 B/op	    1825 allocs/op
BenchmarkContainerImageRender-8          	    2000	    523001 ns/op	  197214 B/op	    2494 allocs/op
BenchmarkContainerImageStats-8           	100000000	        19.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerHostnameRender-8       	    3000	    423759 ns/op	  156442 B/op	    2131 allocs/op
BenchmarkContainerHostnameStats-8        	100000000	        18.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkHostRender-8                    	   10000	    125865 ns/op	   53674 B/op	     591 allocs/op
BenchmarkHostStats-8                     	20000000	        64.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodRender-8                     	    2000	    502690 ns/op	  189855 B/op	    2402 allocs/op
BenchmarkPodStats-8                      	100000000	        19.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodServiceRender-8              	    2000	    596074 ns/op	  226953 B/op	    2728 allocs/op
BenchmarkPodServiceStats-8               	100000000	        18.9 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/weaveworks/scope/render	33.525s
```

master:
```
Toms-MacBook-Pro:render twilkie$ go test -bench .
PASS
BenchmarkEndpointRender-8                	   10000	    136654 ns/op	   52423 B/op	     773 allocs/op
BenchmarkEndpointStats-8                 	100000000	        19.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessRender-8                 	    5000	    252663 ns/op	   97502 B/op	    1372 allocs/op
BenchmarkProcessStats-8                  	20000000	        63.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessWithContainerNameRender-8	    5000	    284485 ns/op	  107775 B/op	    1540 allocs/op
BenchmarkProcessWithContainerNameStats-8 	20000000	        73.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessNameRender-8             	    5000	    322436 ns/op	  121794 B/op	    1639 allocs/op
BenchmarkProcessNameStats-8              	100000000	        19.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerRender-8               	    3000	    438796 ns/op	  167064 B/op	    2390 allocs/op
BenchmarkContainerStats-8                	    3000	    381435 ns/op	  145151 B/op	    2141 allocs/op
BenchmarkContainerWithImageNameRender-8  	    3000	    470924 ns/op	  179896 B/op	    2546 allocs/op
BenchmarkContainerWithImageNameStats-8   	    3000	    388080 ns/op	  145163 B/op	    2142 allocs/op
BenchmarkContainerImageRender-8          	    2000	    559187 ns/op	  218356 B/op	    2908 allocs/op
BenchmarkContainerImageStats-8           	100000000	        19.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerHostnameRender-8       	    3000	    458770 ns/op	  174065 B/op	    2484 allocs/op
BenchmarkContainerHostnameStats-8        	100000000	        19.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkHostRender-8                    	   10000	    145496 ns/op	   61871 B/op	     709 allocs/op
BenchmarkHostStats-8                     	20000000	        69.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodRender-8                     	    2000	    554012 ns/op	  210833 B/op	    2802 allocs/op
BenchmarkPodStats-8                      	100000000	        20.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodServiceRender-8              	    2000	    693266 ns/op	  249112 B/op	    3152 allocs/op
BenchmarkPodServiceStats-8               	50000000	        20.1 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/weaveworks/scope/render	34.110s
```