@paulbellamy 2016-01-20
rebased on master after #752 merge