@paulbellamy 2016-01-20
After rebasing, with a slow report, it looks to give a ~20% decrease in render times.