@tomwilkie 2016-01-20
Sweet!

I think next steps are doing this for metadata (or moving all the uses of
metadata to latest) and doing this for controls.

I also want to figure out a way of measuring the cpu cost on a per path
basis, maybe by just measuring latency of each http request?

On Wednesday, 20 January 2016, Paul Bellamy <notifications@github.com>
wrote:

> After rebasing, with a slow report, it looks to give a ~20% decrease in
> render times.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/weaveworks/scope/pull/838#issuecomment-173175547>.
>