@tomwilkie 2016-01-21
Here's the current performance after rebase:

```
Toms-MacBook-Pro:render twilkie$ go test -bench . -- -bench-report-file report.json 
PASS
BenchmarkEndpointRender-8                	   10000	    119374 ns/op	   46749 B/op	     638 allocs/op
BenchmarkEndpointStats-8                 	100000000	        19.3 ns/op       0 B/op	       0 allocs/op
BenchmarkProcessRender-8                 	    5000	    270373 ns/op	   98960 B/op	    1211 allocs/op
BenchmarkProcessStats-8                  	20000000	        68.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessWithContainerNameRender-8	    5000	    278818 ns/op	  114341 B/op	    1382 allocs/op
BenchmarkProcessWithContainerNameStats-8 	20000000	        76.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessNameRender-8             	    5000	    309158 ns/op	  127276 B/op	    1475 allocs/op
BenchmarkProcessNameStats-8              	100000000	        19.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerRender-8               	    3000	    446979 ns/op	  186119 B/op	    2183 allocs/op
BenchmarkContainerStats-8                	    5000	    371770 ns/op	  153076 B/op	    1944 allocs/op
BenchmarkContainerWithImageNameRender-8  	    3000	    486974 ns/op	  199662 B/op	    2328 allocs/op
BenchmarkContainerWithImageNameStats-8   	    3000	    376005 ns/op	  153085 B/op	    1944 allocs/op
BenchmarkContainerImageRender-8          	    2000	    600634 ns/op	  255532 B/op	    2703 allocs/op
BenchmarkContainerImageStats-8           	100000000	        21.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerHostnameRender-8       	    2000	    579079 ns/op	  193635 B/op	    2277 allocs/op
BenchmarkContainerHostnameStats-8        	100000000	        20.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkHostRender-8                    	    2000	    878810 ns/op	  329441 B/op	    3489 allocs/op
BenchmarkHostStats-8                     	10000000	       139 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodRender-8                     	    2000	    590655 ns/op	  235847 B/op	    2587 allocs/op
BenchmarkPodStats-8                      	100000000	        20.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodServiceRender-8              	    2000	    710204 ns/op	  291872 B/op	    2968 allocs/op
BenchmarkPodServiceStats-8               	100000000	        20.5 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/weaveworks/scope/render	39.250s
```